Update phone field since postgres needs other type

+++ backend/social_network/api/migrations/0001_initial.py
@@ -29,7 +29,7 @@ class Migration(migrations.Migration):
                 ('is_staff', models.BooleanField(default=False, help_text='Designates whether the user can log into this admin site.', verbose_name='staff status')),
                 ('is_active', models.BooleanField(default=True, help_text='Designates whether this user should be treated as active. Unselect this instead of deleting accounts.', verbose_name='active')),
                 ('date_joined', models.DateTimeField(default=django.utils.timezone.now, verbose_name='date joined')),
-                ('phone', models.PositiveIntegerField(blank=True, null=True, unique=True, verbose_name='Номер телефона')),
+                ('phone', models.PositiveBigIntegerField(blank=True, null=True, unique=True, verbose_name='Номер телефона')),
                 ('email', models.EmailField(max_length=254, unique=True, verbose_name='Электронная почта')),
                 ('username', models.CharField(max_length=150, unique=True, validators=[django.core.validators.RegexValidator('^[\\w.@+-]+\\Z', message='Введите допустимое имя')], verbose_name='Уникальный юзернейм')),
                 ('about', models.TextField(blank=True, max_length=500, verbose_name='О себе')),
